audio/avctp: Fix key auto release logic
uinput key codes are 16 bits long not 8 bits, this may cause an overflow
if the key code is above 0xff breaking the auto release logic.

+++ profiles/audio/avctp.c
@@ -178,7 +178,7 @@ struct avctp_channel {
 };
 
 struct key_pressed {
-	uint8_t op;
+	uint16_t op;
 	guint timer;
 };
